fix(analysis): rename types.py → models.py to stop shadowing stdlib
When the analysis package's working directory is on sys.path[0]
(e.g., \`python3 -m analysis\` from inside python/analysis/, or
\`PYTHONPATH=python/analysis ...\`), the local types.py shadows
stdlib's types module. stdlib's threading → functools → from types
import GenericAlias then resolves to analysis.types and fails with
"cannot import name 'MappingProxyType' from partially initialized
module 'types' (most likely due to a circular import)."

This blocks every kernel/observability summary that wants to drive
the analysis lib (today's run hit it when trying to summarize the
day's decision log). The fix is mechanical: rename to a name that
doesn't collide with stdlib.

Renamed: python/analysis/types.py → models.py
Updated 18 importers (loaders / detect / draft / llm_draft / writers
/ templates / 12 test files) from \`from analysis.types import ...\`
to \`from analysis.models import ...\`.

Verification:
- pytest: 142/142 pass (was 142/142 before — same coverage, new name).
- Cold-start import from outside the dir works (PYTHONPATH=python
  python3 -m analysis.decisions --help).
- The classic shadow trap (cwd=python/analysis) no longer triggers.
