+++ python/analysis/templates/no_force_push.py
@@ -7,7 +7,7 @@ from __future__ import annotations
 from typing import Optional
 
 from analysis.templates import register
-from analysis.types import Pattern, PredictedImpact, RuleDraft
+from analysis.models import Pattern, PredictedImpact, RuleDraft
 
 PROTECTED_BRANCHES = frozenset({"main", "master", "production", "release"})
 FEATURE_PREFIXES = ("feat/", "fix/", "spike/", "feature/", "bugfix/", "wip/", "draft/")
